% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/browse_dataset.R
\name{showAllData}
\alias{showAllData}
\title{Show all 35 datasets}
\usage{
showAllData()
}
\value{
A \linkS4class{DataFrame}, containing the following fields
\itemize{
\item \code{Name}, short name of this data.
\item \code{Description}, description of the data.
\item \code{Species}, whether it is human or mouse.
\item \code{SourceUrl}, where the data was originally obtained from.
\item \code{RDataPath}, the corresponding rds file in this package.
}
}
\description{
This shows all the 35 Summary information for all available datasets in the \pkg{scSimBenchmarkData} package.
}
\details{
This package contains a total of 35 curated datasets, covering a range of sequencing protocols, tissues types (including different cell lines) and from both mouse and human.
}
\examples{
showAllData

}
\author{
Yue Cao
}
